Renamed some folders and imports.
Renamed the "/app" folder to "/store". At "/index.js" I imported the "/store/index.js" file.
